Beginning to use logback in the whole system.
When reading the documentation/tutorial of logback I learned how the logback.xml works so logs also get written to /var/log/ or ./logs/privatehome{yyyy-MM-dd}.log depending on whether it runs in portable mode or not.
The log file gets rotated every 24 hours a 0 o'Clock local. And only the last 10 are kept or 3GB what should never be achieved.

The addition of logback.xml lead to akka.http correctly logging bind exceptions so there we only use the shutdown hook to close the Server.

If you need another debug-level just set it in logback.xml (root)