Changed repl to be an object that handles the connection to the console without alpakka.
Because the repl command are not really dependent on the socket and could theoretically could also come from another trusted source.
The parsing and execution of stringCommands is now handled by stringCommandHandler and the interpretMessage methode.

We switched form akka-stream-alpakka-unix-domain-socket to org.scala-sbt.ipcsocket because it is more lightweight and I had a big problem with implementing the two way communication.
So now that we have two way communication we can get a random free ID from the server that way its more easy to add a switch.
Since you do not need to know which id is free or are predictable.