mqtt: Changed the underling structure
Changed form a per switch basis to a structure where mqtt-communication gets handled by mqttController which basically a software representation of the microcontroller on the other side. That is also responsible for the encryption and it supports multiple switches on one controller. (0...63) which on the client get mapped to physical ports/pins.