Added a basic cli
With this version of the cli you can display a help dialog the current
version that gets read from sbt. As real functions you can display the
current status of a switch and also turn one on, off and toggle one with
a specific ID.
If you specify -i and don't specify a valid used ID it will launch the
interactive ID selector.